move color to cli.utils
