+++ brownie/cli/utils/color.py
